+++ test/event_store/support/append_events_test_case.ex
@@ -18,6 +18,33 @@ defmodule Commanded.EventStore.AppendEventsTestCase do
         assert :ok == EventStore.append_to_stream("stream", 4, build_events(1))
       end
 
+      test "should append events without checking expected version" do
+        assert :ok == EventStore.append_to_stream("stream", :any_version, build_events(2))
+        assert :ok == EventStore.append_to_stream("stream", :any_version, build_events(2))
+        assert :ok == EventStore.append_to_stream("stream", :any_version, build_events(1))
+      end
+
+      test "should append events with `:no_stream` parameter" do
+        assert :ok == EventStore.append_to_stream("stream", :no_stream, build_events(2))
+      end
+
+      test "should fail when stream aleady exists with `:no_stream` parameter" do
+        assert :ok == EventStore.append_to_stream("stream", :no_stream, build_events(2))
+
+        assert {:error, :stream_exists} ==
+                 EventStore.append_to_stream("stream", :no_stream, build_events(1))
+      end
+
+      test "should append events with :stream_exists parameter" do
+        assert :ok == EventStore.append_to_stream("stream", :no_stream, build_events(2))
+        assert :ok == EventStore.append_to_stream("stream", :stream_exists, build_events(1))
+      end
+
+      test "should fail when stream does not exists with :stream_exists parameter" do
+        assert {:error, :stream_does_not_exist} ==
+                 EventStore.append_to_stream("stream", :stream_exists, build_events(1))
+      end
+
       test "should fail to append to a stream because of wrong expected version when no previous events" do
         events = build_events(1)