+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Process manager and event handler error & exception handling ([#192](https://github.com/commanded/commanded/pull/192)).
 - Process manager event handling timeout ([#193](https://github.com/commanded/commanded/pull/193)).
 - Allow event handlers to subscribe to individual streams ([#203](https://github.com/commanded/commanded/pull/203)).
+- Add new values for `expected_version` for event store append events behaviour ([#127](https://github.com/commanded/commanded/pull/127)).
 
 ### Bug fixes